requirements.txt
# Core dependencies
google-cloud-firestore==2.13.1
python-dotenv==1.0.0
google-api-python-client==2.92.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0

# Testing dependencies  
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Development dependencies
black==23.11.0
flake8==6.1.0
